src/requirements.txt
amqp==2.5.1
backports.csv==1.0.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-admin-sortable2==0.7.3
django-cleanup==4.0.0
django-colorfield==0.1.15
django-cors-headers==3.1.0
django-debug-toolbar==2.0
django-geoposition-2==0.3.6
django-import-export==1.2.0
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-mptt==0.10.0
django-mptt-admin==0.7.1
django-progressbarupload==0.1.7
django-rest-swagger==2.2.0
git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.10.2
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.6.4
MarkupSafe==1.1.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.3
Pillow==6.1.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.2